requirements.txt
Flask==2.1.2
Pillow==8.3.1
numpy==1.21.2
tensorflow==2.6.0
